Revert "Merge branch 'Printcounter-Fix-(issue-#20785)' of https://github.com/ConstantijnCrijnen/Marlin into Printcounter-Fix-(issue-#20785)"
This reverts commit a5571df42e65e9279fe3f2874cd424f669744ed9, reversing
changes made to 977eaa09ee2fd613846ba044acff3ebd44e4bb1e.
