+++ Marlin/src/HAL/LPC1768/inc/Conditionals_post.h
@@ -26,10 +26,3 @@
 #elif EITHER(I2C_EEPROM, SPI_EEPROM)
   #define USE_SHARED_EEPROM 1
 #endif
-
-// LPC1768 boards seem to lose steps when saving to EEPROM during print (issue #20785)
-// TODO: Which other boards are incompatible?
-#if defined(MCU_LPC1768) && PRINTCOUNTER_SAVE_INTERVAL > 0
-  #warning "To prevent step loss, motion will pause for PRINTCOUNTER auto-save."
-  #define PRINTCOUNTER_SYNC 1
-#endif
