+++ Marlin/src/gcode/control/M80_M81.cpp
@@ -91,7 +91,7 @@ void GcodeSuite::M81() {
   thermalManager.disable_all_heaters();
   planner.finish_and_disable();
 
-  print_job_timer.stop();
+  print_job_timer.stop(); // Wait for planner before calling!
 
   #if HAS_FAN
     thermalManager.zero_fan_speeds();
