+++ Marlin/src/inc/Version.h
@@ -42,7 +42,7 @@
  * version was tagged.
  */
 #ifndef STRING_DISTRIBUTION_DATE
-  #define STRING_DISTRIBUTION_DATE "2021-01-25"
+  #define STRING_DISTRIBUTION_DATE "2021-01-24"
 #endif
 
 /**
