+++ Marlin/src/lcd/dogm/status_screen_DOGM.cpp
@@ -40,10 +40,6 @@
 
 #include "../../gcode/parser.h" // for units (and volumetric)
 
-#if ENABLED(LCD_SHOW_E_TOTAL)
-  #include "../../MarlinCore.h" // for printingIsActive(), marlin_state and MF_SD_COMPLETE
-#endif
-
 #if ENABLED(FILAMENT_LCD_DISPLAY)
   #include "../../feature/filwidth.h"
   #include "../../module/planner.h"
