+++ Marlin/src/lcd/tft/tft_image.h
@@ -21,9 +21,8 @@
  */
 #pragma once
 
+#include "stdint.h"
 #include "../../inc/MarlinConfigPre.h"
-
-#include <stdint.h>
 
 extern const uint8_t marlin_logo_112x38x1[];
 extern const uint8_t marlin_logo_228x255x2[];
