+++ Marlin/src/lcd/tft/ui_480x320.h
@@ -21,29 +21,97 @@
  */
 #pragma once
 
-#define MARLIN_LOGO_FULL_SIZE MarlinLogo480x320x16
+#include "../../inc/MarlinConfigPre.h"
 
-#include "ui_common.h"
+#include "tft.h"
+#include "tft_image.h"
 
-#define TFT_STATUS_TOP_Y       4
-#define TFT_TOP_LINE_Y         4
+#if ENABLED(TOUCH_SCREEN)
+  #include "touch.h"
+#endif
 
-#define MENU_TEXT_X_OFFSET    16
-#define MENU_TEXT_Y_OFFSET     7
+void draw_heater_status(uint16_t x, uint16_t y, const int8_t Heater);
+void draw_fan_status(uint16_t x, uint16_t y, const bool blink);
 
-#define MENU_ITEM_ICON_X       5
-#define MENU_ITEM_ICON_Y       5
-#define MENU_ITEM_ICON_SPACE  42
+#define MENU_TEXT_X_OFFSET  16
+#define MENU_TEXT_Y_OFFSET   7
+void menu_line(const uint8_t row, uint16_t color = COLOR_BACKGROUND);
+void menu_item(const uint8_t row, bool sel = false);
 
 #if HAS_UI_480x320
-  #define MENU_FONT_NAME      Helvetica18
-  #define SYMBOLS_FONT_NAME   Helvetica18_symbols
-  #define MENU_ITEM_HEIGHT    43
-  #define FONT_LINE_HEIGHT    34
+  #define MENU_FONT_NAME    Helvetica18
+  #define SYMBOLS_FONT_NAME Helvetica18_symbols
+  #define MENU_ITEM_HEIGHT  43
+  #define FONT_LINE_HEIGHT  34
 #elif HAS_UI_480x272
-  #define MENU_FONT_NAME      Helvetica14
-  #define SYMBOLS_FONT_NAME   Helvetica14_symbols
-  #define MENU_ITEM_HEIGHT    36
-  #define FONT_LINE_HEIGHT    24
+  #define MENU_FONT_NAME    Helvetica14
+  #define SYMBOLS_FONT_NAME Helvetica14_symbols
+  #define MENU_ITEM_HEIGHT  36
+  #define FONT_LINE_HEIGHT  24
 #endif
-#define MENU_LINE_HEIGHT      (MENU_ITEM_HEIGHT + 2)
+
+#define ABSOLUTE_ZERO     -273.15
+
+const tImage Images[imgCount] = {
+  TERN(SHOW_BOOTSCREEN, TERN(BOOT_MARLIN_LOGO_SMALL, MarlinLogo195x59x16, MarlinLogo480x320x16), NoLogo),
+  HotEnd_64x64x4,
+  Bed_64x64x4,
+  Bed_Heated_64x64x4,
+  Chamber_64x64x4,
+  Chamber_Heated_64x64x4,
+  Fan0_64x64x4,
+  Fan_Slow0_64x64x4,
+  Fan_Slow1_64x64x4,
+  Fan_Fast0_64x64x4,
+  Fan_Fast1_64x64x4,
+  Feedrate_32x32x4,
+  Flowrate_32x32x4,
+  SD_64x64x4,
+  Menu_64x64x4,
+  Settings_64x64x4,
+  Directory_32x32x4,
+  Confirm_64x64x4,
+  Cancel_64x64x4,
+  Increase_64x64x4,
+  Decrease_64x64x4,
+  Back_32x32x4,
+  Up_32x32x4,
+  Down_32x32x4,
+  Left_32x32x4,
+  Right_32x32x4,
+  Refresh_32x32x4,
+  Leveling_32x32x4,
+  Slider8x16x4,
+  Home_64x64x4,
+  BtnRounded_64x52x4,
+};
+
+#if HAS_TEMP_CHAMBER && HOTENDS > 1
+  #define ITEM_E0         0
+  #define ITEM_E1         1
+  #define ITEM_BED        2
+  #define ITEM_CHAMBER    3
+  #define ITEM_FAN        4
+  #define ITEMS_COUNT     5
+  #define POS_Y           4
+#elif HAS_TEMP_CHAMBER
+  #define ITEM_E0         0
+  #define ITEM_BED        1
+  #define ITEM_CHAMBER    2
+  #define ITEM_FAN        3
+  #define ITEMS_COUNT     4
+  #define POS_Y           4
+#elif HOTENDS > 1
+  #define ITEM_E0         0
+  #define ITEM_E1         1
+  #define ITEM_BED        2
+  #define ITEM_FAN        3
+  #define ITEMS_COUNT     4
+  #define POS_Y           4
+#else
+  #define ITEM_E0         0
+  #define ITEM_BED        1
+  #define ITEM_FAN        2
+  #define ITEMS_COUNT     3
+  #define POS_Y           4
+#endif
