+++ Marlin/src/module/printcounter.h
@@ -71,18 +71,19 @@ class PrintCounter: public Stopwatch {
      * @brief Interval in seconds between counter updates
      * @details This const value defines what will be the time between each
      * accumulator update. This is different from the EEPROM save interval.
+     *
+     * @note The max value for this option is 60(s), otherwise integer
+     * overflow will happen.
      */
-    static constexpr millis_t updateInterval = SEC_TO_MS(10);
+    static constexpr uint16_t updateInterval = 10;
 
-    #if PRINTCOUNTER_SAVE_INTERVAL > 0
-      /**
-       * @brief Interval in seconds between EEPROM saves
-       * @details This const value defines what will be the time between each
-       * EEPROM save cycle, the development team recommends to set this value
-       * no lower than 3600 secs (1 hour).
-       */
-      static constexpr millis_t saveInterval = MIN_TO_MS(PRINTCOUNTER_SAVE_INTERVAL);
-    #endif
+    /**
+     * @brief Interval in seconds between EEPROM saves
+     * @details This const value defines what will be the time between each
+     * EEPROM save cycle, the development team recommends to set this value
+     * no lower than 3600 secs (1 hour).
+     */
+    static constexpr uint16_t saveInterval = 3600;
 
     /**
      * @brief Timestamp of the last call to deltaDuration()
